Extract server request normalization
